+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # TelegramBotAPI changelog
 
+## 0.10.0
 
 ## 0.9.0